Starting to clean up MainActivity2
