+++ app/src/main/java/com/samourai/wallet/MainActivity2.java
@@ -11,30 +11,17 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.pm.ActivityInfo;
 import android.os.Bundle;
-import android.os.Environment;
-import android.os.Handler;
 import android.os.Looper;
 import android.support.v4.content.LocalBroadcastManager;
-import android.text.Editable;
-import android.text.InputType;
-import android.text.TextWatcher;
-import android.view.Gravity;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
+import android.util.Log;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
-import android.widget.LinearLayout;
-import android.widget.Toast;
-//import android.util.Log;
-
-import org.bitcoinj.crypto.MnemonicException;
 
 import com.samourai.wallet.access.AccessFactory;
 import com.samourai.wallet.api.APIFactory;
 import com.samourai.wallet.crypto.AESUtil;
-import com.samourai.wallet.crypto.DecryptionException;
-import com.samourai.wallet.hd.HD_Wallet;
-import com.samourai.wallet.hd.HD_WalletFactory;
+import com.samourai.wallet.initialization.ExchangeRateThread;
+import com.samourai.wallet.initialization.NewWalletDialog;
+import com.samourai.wallet.initialization.RestoreWalletDialog;
 import com.samourai.wallet.payload.PayloadUtil;
 import com.samourai.wallet.prng.PRNGFixes;
 import com.samourai.wallet.service.BroadcastReceiverService;
@@ -42,27 +29,18 @@ import com.samourai.wallet.service.WebSocketService;
 import com.samourai.wallet.util.AppUtil;
 import com.samourai.wallet.util.CharSequenceX;
 import com.samourai.wallet.util.ConnectivityStatus;
-import com.samourai.wallet.util.ExchangeRateFactory;
+import com.samourai.wallet.util.LogUtil;
 import com.samourai.wallet.util.PrefsUtil;
 import com.samourai.wallet.util.TimeOutUtil;
-import com.samourai.wallet.util.WebUtil;
 
 import org.apache.commons.codec.DecoderException;
+import org.bitcoinj.crypto.MnemonicException;
 import org.bitcoinj.params.MainNetParams;
-import org.bitcoinj.params.TestNet2Params;
 import org.bitcoinj.params.TestNet3Params;
-import org.json.JSONException;
-import org.json.JSONObject;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.regex.Pattern;
 
 public class MainActivity2 extends Activity {
+    private static final String TAG = LogUtil.getTag();
 
     private ProgressDialog progress = null;
 
@@ -98,14 +76,14 @@ public class MainActivity2 extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        if (LogUtil.DEBUG) Log.d(TAG, "onCreate");
         setContentView(R.layout.activity_main);
 
         setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
 
         loadedBalanceFragment = false;
 
-//        doAccountSelection();
-
+        if (LogUtil.DEBUG) Log.d(TAG, "setup action bar");
         getActionBar().setNavigationMode(ActionBar.NAVIGATION_MODE_LIST);
         ActionBar.OnNavigationListener navigationListener = new ActionBar.OnNavigationListener() {
             @Override
@@ -152,9 +130,12 @@ public class MainActivity2 extends Activity {
             AppUtil.getInstance(MainActivity2.this).setPRNG_FIXED(true);
         }
 
-        if(!ConnectivityStatus.hasConnectivity(MainActivity2.this) &&
-        !(AccessFactory.getInstance(MainActivity2.this).getGUID().length() < 1 || !PayloadUtil.getInstance(MainActivity2.this).walletFileExists())) {
+        //check if no internetz
+        if(!ConnectivityStatus.hasConnectivity(MainActivity2.this)
+                && !(AccessFactory.getInstance(MainActivity2.this).getGUID().length() < 1
+                || !PayloadUtil.getInstance(MainActivity2.this).walletFileExists())) {
 
+            if (LogUtil.DEBUG) Log.d(TAG, "no internet connection");
             new AlertDialog.Builder(MainActivity2.this)
                     .setTitle(R.string.app_name)
                     .setMessage(R.string.no_internet)
@@ -166,9 +147,8 @@ public class MainActivity2 extends Activity {
                     }).show();
 
         }
+        //there is internetz
         else  {
-//            SSLVerifierThreadUtil.getInstance(MainActivity2.this).validateSSLThread();
-//            APIFactory.getInstance(MainActivity2.this).validateAPIThread();
             exchangeRateThread();
 
             boolean isDial = false;
@@ -195,33 +175,9 @@ public class MainActivity2 extends Activity {
     protected void onResume() {
         super.onResume();
 
-        AppUtil.getInstance(MainActivity2.this).setIsInForeground(true);
-
         AppUtil.getInstance(MainActivity2.this).deleteQR();
         AppUtil.getInstance(MainActivity2.this).deleteBackup();
-/*
-        if(TimeOutUtil.getInstance().isTimedOut()) {
-            if(AccessFactory.getInstance(MainActivity2.this).getGUID().length() < 1 || !PayloadUtil.getInstance(MainActivity2.this).walletFileExists()) {
-                AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
-                if(AppUtil.getInstance(MainActivity2.this).isSideLoaded())    {
-                    doSelectNet();
-                }
-                else    {
-                    initDialog();
-                }
-            }
-            else {
-                AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
-                validatePIN(null);
-            }
-        }
-        else {
-            TimeOutUtil.getInstance().updatePin();
 
-//            SSLVerifierThreadUtil.getInstance(MainActivity2.this).validateSSLThread();
-//            APIFactory.getInstance(MainActivity2.this).validateAPIThread();
-        }
-*/
         IntentFilter filter_restart = new IntentFilter(ACTION_RESTART);
         LocalBroadcastManager.getInstance(MainActivity2.this).registerReceiver(receiver_restart, filter_restart);
 
@@ -232,10 +188,7 @@ public class MainActivity2 extends Activity {
     @Override
     protected void onPause() {
         super.onPause();
-
         LocalBroadcastManager.getInstance(MainActivity2.this).unregisterReceiver(receiver_restart);
-
-        AppUtil.getInstance(MainActivity2.this).setIsInForeground(false);
     }
 
     @Override
@@ -248,358 +201,30 @@ public class MainActivity2 extends Activity {
     }
 
     private void initDialog()	{
-
+        if (LogUtil.DEBUG) Log.d(TAG, "initDialog");
         AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
 
+        //Chose create new or restore old wallet
         AlertDialog.Builder dlg = new AlertDialog.Builder(this)
                 .setTitle(R.string.app_name)
                 .setMessage(R.string.please_select)
                 .setCancelable(false)
                 .setPositiveButton(R.string.create_wallet, new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int whichButton) {
+                        if (LogUtil.DEBUG) Log.d(TAG, "createWallet");
 
-                        final EditText passphrase = new EditText(MainActivity2.this);
-                        passphrase.setSingleLine(true);
-                        passphrase.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS);
-
-                        AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                .setTitle(R.string.app_name)
-                                .setMessage(R.string.bip39_safe)
-                                .setView(passphrase)
-                                .setCancelable(false)
-                                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                        final String passphrase39 = passphrase.getText().toString();
-
-                                        if(passphrase39 != null && passphrase39.length() > 0 && passphrase39.contains(" "))    {
-
-                                            Toast.makeText(MainActivity2.this, R.string.bip39_invalid, Toast.LENGTH_SHORT).show();
-                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                        }
-                                        else if (passphrase39 != null && passphrase39.length() > 0) {
-
-                                            final EditText passphrase2 = new EditText(MainActivity2.this);
-                                            passphrase2.setSingleLine(true);
-                                            passphrase2.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS);
-
-                                            AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                                    .setTitle(R.string.app_name)
-                                                    .setMessage(R.string.bip39_safe2)
-                                                    .setView(passphrase2)
-                                                    .setCancelable(false)
-                                                    .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                                                        public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                            final String _passphrase39 = passphrase2.getText().toString();
-
-                                                            if(_passphrase39.equals(passphrase39))    {
-
-                                                                Intent intent = new Intent(MainActivity2.this, PinEntryActivity.class);
-                                                                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                                                                intent.putExtra("create", true);
-                                                                intent.putExtra("passphrase", _passphrase39);
-                                                                startActivity(intent);
-
-                                                            }
-                                                            else {
-
-                                                                Toast.makeText(MainActivity2.this, R.string.bip39_unmatch, Toast.LENGTH_SHORT).show();
-                                                                AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                            }
-
-                                                        }
-
-                                                    }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                                                        public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                            Toast.makeText(MainActivity2.this, R.string.bip39_must, Toast.LENGTH_SHORT).show();
-                                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                        }
-                                                    });
-                                            if(!isFinishing())    {
-                                                dlg.show();
-                                            }
-
-                                        }
-                                        else {
-
-                                            Toast.makeText(MainActivity2.this, R.string.bip39_must, Toast.LENGTH_SHORT).show();
-                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                        }
-
-                                    }
-
-                                }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                        Toast.makeText(MainActivity2.this, R.string.bip39_must, Toast.LENGTH_SHORT).show();
-                                        AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                    }
-                                });
                         if(!isFinishing())    {
-                            dlg.show();
+                            NewWalletDialog.makeDialog(MainActivity2.this).show();
                         }
-
                     }
                 })
                 .setNegativeButton(R.string.restore_wallet, new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int whichButton) {
+                        if (LogUtil.DEBUG) Log.d(TAG, "restoreWallet");
 
-                        AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                .setTitle(R.string.app_name)
-                                .setMessage(R.string.restore_wallet)
-                                .setCancelable(false)
-                                .setPositiveButton(R.string.import_backup, new DialogInterface.OnClickListener() {
-                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                        final EditText passphrase = new EditText(MainActivity2.this);
-                                        passphrase.setInputType(InputType.TYPE_CLASS_TEXT | InputType.TYPE_TEXT_VARIATION_PASSWORD | InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS);
-                                        passphrase.setHint(R.string.passphrase);
-
-                                        AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                                .setTitle(R.string.app_name)
-                                                .setView(passphrase)
-                                                .setMessage(R.string.restore_wallet_from_backup)
-                                                .setCancelable(false)
-                                                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                        final String pw = passphrase.getText().toString();
-                                                        if (pw == null || pw.length() < 1) {
-                                                            Toast.makeText(MainActivity2.this, R.string.invalid_passphrase, Toast.LENGTH_SHORT).show();
-                                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-                                                        }
-
-                                                        String data = null;
-                                                        File file = PayloadUtil.getInstance(MainActivity2.this).getBackupFile();
-                                                        if(file != null && file.exists())    {
-
-                                                            StringBuilder sb = new StringBuilder();
-                                                            try {
-                                                                BufferedReader in = new BufferedReader(new InputStreamReader(new FileInputStream(file), "UTF8"));
-                                                                String str = null;
-                                                                while((str = in.readLine()) != null) {
-                                                                    sb.append(str);
-                                                                }
-
-                                                                in.close();
-                                                            }
-                                                            catch(FileNotFoundException fnfe) {
-
-                                                            }
-                                                            catch(IOException ioe) {
-
-                                                            }
-
-                                                            data = sb.toString();
-
-                                                        }
-
-                                                        final EditText edBackup = new EditText(MainActivity2.this);
-                                                        edBackup.setSingleLine(false);
-                                                        edBackup.setInputType(InputType.TYPE_CLASS_TEXT | InputType.TYPE_TEXT_FLAG_MULTI_LINE);
-                                                        edBackup.setLines(7);
-                                                        edBackup.setHint(R.string.encrypted_backup);
-                                                        edBackup.setGravity(Gravity.START);
-                                                        TextWatcher textWatcher = new TextWatcher() {
-
-                                                            public void afterTextChanged(Editable s) {
-                                                                edBackup.setSelection(0);
-                                                            }
-                                                            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                                                                ;
-                                                            }
-                                                            public void onTextChanged(CharSequence s, int start, int before, int count) {
-                                                                ;
-                                                            }
-                                                        };
-                                                        edBackup.addTextChangedListener(textWatcher);
-                                                        String message = null;
-                                                        if(data != null)   {
-                                                            edBackup.setText(data);
-                                                            message = getText(R.string.restore_wallet_from_existing_backup).toString();
-                                                        }
-                                                        else    {
-                                                            message = getText(R.string.restore_wallet_from_backup).toString();
-                                                        }
-
-                                                        AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                                                .setTitle(R.string.app_name)
-                                                                .setView(edBackup)
-                                                                .setMessage(message)
-                                                                .setCancelable(false)
-                                                                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                                        String data = edBackup.getText().toString();
-                                                                        if (data == null || data.length() < 1) {
-                                                                            Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-                                                                        }
-
-                                                                        final String decrypted = PayloadUtil.getInstance(MainActivity2.this).getDecryptedBackupPayload(data, new CharSequenceX(pw));
-                                                                        if(decrypted == null || decrypted.length() < 1)    {
-                                                                            Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-                                                                        }
-
-                                                                        if (progress != null && progress.isShowing()) {
-                                                                            progress.dismiss();
-                                                                            progress = null;
-                                                                        }
-
-                                                                        progress = new ProgressDialog(MainActivity2.this);
-                                                                        progress.setCancelable(false);
-                                                                        progress.setTitle(R.string.app_name);
-                                                                        progress.setMessage(getString(R.string.please_wait));
-                                                                        progress.show();
-
-                                                                        new Thread(new Runnable() {
-                                                                            @Override
-                                                                            public void run() {
-                                                                                Looper.prepare();
-
-                                                                                try {
-
-                                                                                    JSONObject json = new JSONObject(decrypted);
-                                                                                    HD_Wallet hdw = PayloadUtil.getInstance(MainActivity2.this).restoreWalletfromJSON(json);
-                                                                                    HD_WalletFactory.getInstance(MainActivity2.this).set(hdw);
-                                                                                    String guid = AccessFactory.getInstance(MainActivity2.this).createGUID();
-                                                                                    String hash = AccessFactory.getInstance(MainActivity2.this).getHash(guid, new CharSequenceX(AccessFactory.getInstance(MainActivity2.this).getPIN()), AESUtil.DefaultPBKDF2Iterations);
-                                                                                    PrefsUtil.getInstance(MainActivity2.this).setValue(PrefsUtil.ACCESS_HASH, hash);
-                                                                                    PrefsUtil.getInstance(MainActivity2.this).setValue(PrefsUtil.ACCESS_HASH2, hash);
-                                                                                    PayloadUtil.getInstance(MainActivity2.this).saveWalletToJSON(new CharSequenceX(guid + AccessFactory.getInstance().getPIN()));
-
-                                                                                }
-                                                                                catch(MnemonicException.MnemonicLengthException mle) {
-                                                                                    mle.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                catch(DecoderException de) {
-                                                                                    de.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                catch(JSONException je) {
-                                                                                    je.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                catch(IOException ioe) {
-                                                                                    ioe.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                catch(java.lang.NullPointerException npe) {
-                                                                                    npe.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                catch(DecryptionException de) {
-                                                                                    de.printStackTrace();
-                                                                                    Toast.makeText(MainActivity2.this, R.string.decryption_error, Toast.LENGTH_SHORT).show();
-                                                                                }
-                                                                                finally {
-                                                                                    if (progress != null && progress.isShowing()) {
-                                                                                        progress.dismiss();
-                                                                                        progress = null;
-                                                                                    }
-                                                                                    AppUtil.getInstance(MainActivity2.this).restartApp();
-                                                                                }
-
-                                                                                Looper.loop();
-
-                                                                            }
-                                                                        }).start();
-
-                                                                    }
-                                                                }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                                        AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                                    }
-                                                                });
-                                                        if(!isFinishing())    {
-                                                            dlg.show();
-                                                        }
-
-                                                    }
-                                                }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                        AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                    }
-                                                });
-                                        if(!isFinishing())    {
-                                            dlg.show();
-                                        }
-
-                                    }
-                                })
-                                .setNegativeButton(R.string.import_mnemonic, new DialogInterface.OnClickListener() {
-                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                        final EditText mnemonic = new EditText(MainActivity2.this);
-                                        mnemonic.setHint(R.string.mnemonic_hex);
-                                        mnemonic.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS);
-                                        final EditText passphrase = new EditText(MainActivity2.this);
-                                        passphrase.setHint(R.string.bip39_passphrase);
-                                        passphrase.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS);
-                                        passphrase.setSingleLine(true);
-
-                                        LinearLayout restoreLayout = new LinearLayout(MainActivity2.this);
-                                        restoreLayout.setOrientation(LinearLayout.VERTICAL);
-                                        restoreLayout.addView(mnemonic);
-                                        restoreLayout.addView(passphrase);
-
-                                        AlertDialog.Builder dlg = new AlertDialog.Builder(MainActivity2.this)
-                                                .setTitle(R.string.app_name)
-                                                .setMessage(R.string.bip39_restore)
-                                                .setView(restoreLayout)
-                                                .setCancelable(false)
-                                                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                        PrefsUtil.getInstance(MainActivity2.this).setValue(PrefsUtil.IS_RESTORE, true);
-
-                                                        final String seed39 = mnemonic.getText().toString();
-                                                        final String passphrase39 = passphrase.getText().toString();
-
-                                                        if (seed39 != null && seed39.length() > 0) {
-                                                            Intent intent = new Intent(MainActivity2.this, PinEntryActivity.class);
-                                                            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                                                            intent.putExtra("create", true);
-                                                            intent.putExtra("seed", seed39);
-                                                            intent.putExtra("passphrase", passphrase39 == null ? "" : passphrase39);
-                                                            startActivity(intent);
-                                                        } else {
-
-                                                            AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                        }
-
-                                                    }
-                                                }).setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-                                                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                                                        AppUtil.getInstance(MainActivity2.this).restartApp();
-
-                                                    }
-                                                });
-                                        if(!isFinishing())    {
-                                            dlg.show();
-                                        }
-
-                                    }
-                                });
                         if(!isFinishing())    {
-                            dlg.show();
+                            RestoreWalletDialog.makeDialog(MainActivity2.this).show();
                         }
-
                     }
                 });
         if(!isFinishing())    {
@@ -656,13 +281,9 @@ public class MainActivity2 extends Activity {
                     TimeOutUtil.getInstance().updatePin();
                     AppUtil.getInstance(MainActivity2.this).restartApp();
                 }
-                catch (MnemonicException.MnemonicLengthException mle) {
+                catch (MnemonicException.MnemonicLengthException | DecoderException mle) {
                     mle.printStackTrace();
-                }
-                catch (DecoderException de) {
-                    de.printStackTrace();
-                }
-                finally {
+                } finally {
                     if (progress != null && progress.isShowing()) {
                         progress.dismiss();
                         progress = null;
@@ -677,81 +298,19 @@ public class MainActivity2 extends Activity {
     }
 
     private void exchangeRateThread() {
-
-        final Handler handler = new Handler();
-
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                Looper.prepare();
-
-                String response = null;
-                try {
-                    response = WebUtil.getInstance(null).getURL(WebUtil.LBC_EXCHANGE_URL);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).setDataLBC(response);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).parseLBC();
-                }
-                catch(Exception e) {
-                    e.printStackTrace();
-                }
-
-                response = null;
-                try {
-                    response = WebUtil.getInstance(null).getURL(WebUtil.BTCe_EXCHANGE_URL + "btc_usd");
-                    ExchangeRateFactory.getInstance(MainActivity2.this).setDataBTCe(response);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).parseBTCe();
-                }
-                catch(Exception e) {
-                    e.printStackTrace();
-                }
-
-                response = null;
-                try {
-                    response = WebUtil.getInstance(null).getURL(WebUtil.BTCe_EXCHANGE_URL + "btc_rur");
-                    ExchangeRateFactory.getInstance(MainActivity2.this).setDataBTCe(response);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).parseBTCe();
-                }
-                catch(Exception e) {
-                    e.printStackTrace();
-                }
-
-                response = null;
-                try {
-                    response = WebUtil.getInstance(null).getURL(WebUtil.BTCe_EXCHANGE_URL + "btc_eur");
-                    ExchangeRateFactory.getInstance(MainActivity2.this).setDataBTCe(response);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).parseBTCe();
-                }
-                catch(Exception e) {
-                    e.printStackTrace();
-                }
-
-                response = null;
-                try {
-                    response = WebUtil.getInstance(null).getURL(WebUtil.BFX_EXCHANGE_URL);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).setDataBFX(response);
-                    ExchangeRateFactory.getInstance(MainActivity2.this).parseBFX();
-                }
-                catch(Exception e) {
-                    e.printStackTrace();
-                }
-
-                handler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        ;
-                    }
-                });
-
-                Looper.loop();
-
-            }
-        }).start();
+        ExchangeRateThread exchangeThread = new ExchangeRateThread(MainActivity2.this);
+        exchangeThread.startExchangeRateCheck();
     }
 
     private void doAppInit(boolean isDial, final String strUri, final String strPCode) {
+        if (LogUtil.DEBUG) Log.d(TAG, "doAppInit -"
+                + " isDial: " + isDial
+                + " strUri: " + strUri
+                + " strPCode: " + strPCode);
 
-        if((strUri != null || strPCode != null) && AccessFactory.getInstance(MainActivity2.this).isLoggedIn())    {
-
+        //IF 1 -
+        if((strUri != null || strPCode != null) && AccessFactory.getInstance(MainActivity2.this).isLoggedIn()) {
+            if (LogUtil.DEBUG) Log.d(TAG, "IF 1");
             progress = new ProgressDialog(MainActivity2.this);
             progress.setCancelable(false);
             progress.setTitle(R.string.app_name);
@@ -781,7 +340,10 @@ public class MainActivity2 extends Activity {
             }).start();
 
         }
+        //IF 2 -
         else if(AccessFactory.getInstance(MainActivity2.this).getGUID().length() < 1 || !PayloadUtil.getInstance(MainActivity2.this).walletFileExists()) {
+            if (LogUtil.DEBUG) Log.d(TAG, "IF 2");
+
             AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
             if(AppUtil.getInstance(MainActivity2.this).isSideLoaded())    {
                 doSelectNet();
@@ -790,16 +352,21 @@ public class MainActivity2 extends Activity {
                 initDialog();
             }
         }
+        //IF 3 -
         else if(isDial && AccessFactory.getInstance(MainActivity2.this).validateHash(PrefsUtil.getInstance(MainActivity2.this).getValue(PrefsUtil.ACCESS_HASH, ""), AccessFactory.getInstance(MainActivity2.this).getGUID(), new CharSequenceX(AccessFactory.getInstance(MainActivity2.this).getPIN()), AESUtil.DefaultPBKDF2Iterations)) {
+            if (LogUtil.DEBUG) Log.d(TAG, "IF 3");
             TimeOutUtil.getInstance().updatePin();
             launchFromDialer(AccessFactory.getInstance(MainActivity2.this).getPIN());
         }
+        //IF 4 -
         else if(TimeOutUtil.getInstance().isTimedOut()) {
+            if (LogUtil.DEBUG) Log.d(TAG, "IF 4");
             AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
-            validatePIN(strUri == null ? null : strUri);
+            validatePIN(strUri);
         }
+        //IF 5 -
         else if(AccessFactory.getInstance(MainActivity2.this).isLoggedIn() && !TimeOutUtil.getInstance().isTimedOut()) {
-
+            if (LogUtil.DEBUG) Log.d(TAG, "IF 5");
             TimeOutUtil.getInstance().updatePin();
             loadedBalanceFragment = true;
 
@@ -808,9 +375,11 @@ public class MainActivity2 extends Activity {
             intent.putExtra("fetch", true);
             startActivity(intent);
         }
+        //ELSE
         else {
+            if (LogUtil.DEBUG) Log.d(TAG, "ELSE");
             AccessFactory.getInstance(MainActivity2.this).setIsLoggedIn(false);
-            validatePIN(strUri == null ? null : strUri);
+            validatePIN(strUri);
         }
 
     }
@@ -821,28 +390,13 @@ public class MainActivity2 extends Activity {
             return;
         }
 
-        /*
-        if(AddressFactory.getInstance(MainActivity2.this).getHighestTxReceiveIdx(SamouraiWallet.MIXING_ACCOUNT) < 1)    {
-            account_selections = new String[] {
-                    getString(R.string.account_samourai),
-            };
-        }
-        else    {
-            account_selections = new String[] {
-                    getString(R.string.total),
-                    getString(R.string.account_samourai),
-                    getString(R.string.account_shuffling),
-            };
-        }
-        */
-
         account_selections = new String[] {
                 getString(R.string.total),
                 getString(R.string.account_Samourai),
                 getString(R.string.account_shuffling),
         };
 
-        adapter = new ArrayAdapter<String>(getBaseContext(), android.R.layout.simple_spinner_dropdown_item, account_selections);
+        adapter = new ArrayAdapter<>(getBaseContext(), android.R.layout.simple_spinner_dropdown_item, account_selections);
 
         if(account_selections.length > 1)    {
             SamouraiWallet.getInstance().setShowTotalBalance(true);
@@ -854,7 +408,7 @@ public class MainActivity2 extends Activity {
     }
 
     private void doSelectNet()  {
-
+        if (LogUtil.DEBUG) Log.d(TAG, "doSelectNet");
         AlertDialog.Builder dlg = new AlertDialog.Builder(this)
                 .setTitle(R.string.app_name)
                 .setMessage(R.string.select_network)
